Network.AWS.Kinesis

Description

Amazon Kinesis is a managed service that scales elastically for real-time processing of streaming big data. The service takes in large streams of data records that can then be consumed in real time by multiple data-processing applications that can be run on Amazon EC2 instances.

Documentation

module Network.AWS.Kinesis.AddTagsToStream

module Network.AWS.Kinesis.CreateStream

module Network.AWS.Kinesis.DeleteStream

module Network.AWS.Kinesis.DescribeStream

module Network.AWS.Kinesis.GetRecords

module Network.AWS.Kinesis.GetShardIterator

module Network.AWS.Kinesis.ListStreams

module Network.AWS.Kinesis.ListTagsForStream

module Network.AWS.Kinesis.MergeShards

module Network.AWS.Kinesis.PutRecord

module Network.AWS.Kinesis.PutRecords

module Network.AWS.Kinesis.RemoveTagsFromStream

module Network.AWS.Kinesis.SplitShard

module Network.AWS.Kinesis.Types

module Network.AWS.Kinesis.Waiters