Network.AWS.Kinesis.DecreaseStreamRetentionPeriod

Description

Decreases the Amazon Kinesis stream's retention period, which is the length of time data records are accessible after they are added to the stream. The minimum value of a stream's retention period is 24 hours.

This operation may result in lost data. For example, if the stream's retention period is 48 hours and is decreased to 24 hours, any data already in the stream that is older than 24 hours is inaccessible.

Creating a Request

decreaseStreamRetentionPeriod

Arguments

:: Text
dsrpStreamName

-> Natural
dsrpRetentionPeriodHours

-> DecreaseStreamRetentionPeriod

Creates a value of DecreaseStreamRetentionPeriod with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrpStreamName - The name of the stream to modify.

dsrpRetentionPeriodHours - The new retention period of the stream, in hours. Must be less than the current retention period.

data DecreaseStreamRetentionPeriod

Represents the input for DecreaseStreamRetentionPeriod .

See: decreaseStreamRetentionPeriod smart constructor.

Request Lenses

dsrpStreamName :: Lens' DecreaseStreamRetentionPeriod Text

The name of the stream to modify.

dsrpRetentionPeriodHours :: Lens' DecreaseStreamRetentionPeriod Natural

The new retention period of the stream, in hours. Must be less than the current retention period.

Destructuring the Response

decreaseStreamRetentionPeriodResponse :: DecreaseStreamRetentionPeriodResponse

Creates a value of DecreaseStreamRetentionPeriodResponse with the minimum fields required to make a request.

data DecreaseStreamRetentionPeriodResponse

See: decreaseStreamRetentionPeriodResponse smart constructor.