Network.AWS.Kinesis.DeleteStream

Description

Deletes a stream and all its shards and data. You must shut down any applications that are operating on the stream before you delete the stream. If an application attempts to operate on a deleted stream, it will receive the exception ResourceNotFoundException. If the stream is in the ACTIVE state, you can delete it. After a DeleteStream request, the specified stream is in the DELETING state until Amazon Kinesis completes the deletion. Note: Amazon Kinesis might continue to accept data read and write operations, such as PutRecord and GetRecords, on a stream in the DELETING state until the stream deletion is complete. When you delete a stream, any shards in that stream are also deleted, and any tags are dissociated from the stream. You can use the DescribeStream operation to check the state of the stream, which is returned in StreamStatus. DeleteStream has a limit of 5 transactions per second per account. To delete a stream The following example deletes the specified stream. POST HTTP1.1 Host: kinesis.. x-amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Kinesis_20131202.DeleteStream { StreamName:"exampleStreamName" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]>.

Request

data DeleteStream

Represents the input for DeleteStream.

Request constructor

deleteStream

Arguments

:: Text
dsStreamName

-> DeleteStream

Smart constructor for the minimum required parameters to construct a valid DeleteStream request.

The fields accessible through corresponding lenses are:

StreamName :: Text

Request lenses

dsStreamName :: Lens' DeleteStream Text

The name of the stream to delete.

Response

data DeleteStreamResponse

Response constructor

deleteStreamResponse :: DeleteStreamResponse

Smart constructor for the minimum required parameters to construct a valid DeleteStreamResponse response.

This constructor is provided for convenience and testing purposes.