Network.AWS.Kinesis.DescribeLimits

Description

Describes the shard limits and usage for the account.

If you update your account limits, the old limits might be returned for a few minutes.

This operation has a limit of 1 transaction per second per account.

Creating a Request

describeLimits :: DescribeLimits

Creates a value of DescribeLimits with the minimum fields required to make a request.

data DescribeLimits

See: describeLimits smart constructor.

Destructuring the Response

describeLimitsResponse

Arguments

:: Int
dlrsResponseStatus

-> Natural
dlrsShardLimit

-> Natural
dlrsOpenShardCount

-> DescribeLimitsResponse

Creates a value of DescribeLimitsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlrsResponseStatus - -- | The response status code.

dlrsShardLimit - The maximum number of shards.

dlrsOpenShardCount - The number of open shards.

data DescribeLimitsResponse

See: describeLimitsResponse smart constructor.

Response Lenses

dlrsResponseStatus :: Lens' DescribeLimitsResponse Int

- | The response status code.

dlrsShardLimit :: Lens' DescribeLimitsResponse Natural

The maximum number of shards.

dlrsOpenShardCount :: Lens' DescribeLimitsResponse Natural

The number of open shards.