Network.AWS.Kinesis.ListTagsForStream

Description

Lists the tags for the specified Amazon Kinesis stream.

http://docs.aws.amazon.com/kinesis/latest/APIReference/API_ListTagsForStream.html

Request

data ListTagsForStream

Request constructor

listTagsForStream

Arguments

:: Text
ltfsStreamName

-> ListTagsForStream

ListTagsForStream constructor.

The fields accessible through corresponding lenses are:

ltfsExclusiveStartTagKey :: Maybe Text

ltfsLimit :: Maybe Natural

ltfsStreamName :: Text

Request lenses

ltfsExclusiveStartTagKey :: Lens' ListTagsForStream (Maybe Text)

The key to use as the starting point for the list of tags. If this parameter is set, ListTagsForStream gets all tags that occur after ExclusiveStartTagKey.

ltfsLimit :: Lens' ListTagsForStream (Maybe Natural)

The number of tags to return. If this number is less than the total number of tags associated with the stream, HasMoreTags is set to true. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

ltfsStreamName :: Lens' ListTagsForStream Text

The name of the stream.

Response

data ListTagsForStreamResponse

Response constructor

listTagsForStreamResponse

Arguments

:: Bool
ltfsrHasMoreTags

-> ListTagsForStreamResponse

ListTagsForStreamResponse constructor.

The fields accessible through corresponding lenses are:

ltfsrHasMoreTags :: Bool

ltfsrTags :: [Tag]

Response lenses

ltfsrHasMoreTags :: Lens' ListTagsForStreamResponse Bool

If set to true, more tags are available. To request additional tags, set ExclusiveStartTagKey to the key of the last tag returned.

ltfsrTags :: Lens' ListTagsForStreamResponse [Tag]

A list of tags associated with StreamName, starting with the first tag after ExclusiveStartTagKey and up to the specified Limit.