Network.AWS.Kinesis.PutRecord

Description

Puts a data record from a producer into an Amazon Kinesis stream. You must call PutRecord to send data from the producer into the Amazon Kinesis stream for real-time ingestion and subsequent processing. You must specify the name of the stream that captures, stores, and transports the data; a partition key; and the data blob itself. The data blob could be a segment from a log file, geographic/location data, website clickstream data, or any other data type. The partition key is used to distribute data across shards. Amazon Kinesis segregates the data records that belong to a data stream into multiple shards, using the partition key associated with each data record to determine which shard a given data record belongs to. Partition keys are Unicode strings, with a maximum length limit of 256 bytes. An MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards using the hash key ranges of the shards. You can override hashing the partition key to determine the shard by explicitly specifying a hash value using the ExplicitHashKey parameter. For more information, see the Amazon Kinesis Developer Guide. PutRecord returns the shard ID of where the data record was placed and the sequence number that was assigned to the data record. Sequence numbers generally increase over time. To guarantee strictly increasing ordering, use the SequenceNumberForOrdering parameter. For more information, see the Amazon Kinesis Developer Guide. If a PutRecord request cannot be processed because of insufficient provisioned throughput on the shard involved in the request, PutRecord throws ProvisionedThroughputExceededException. Data records are accessible for only 24 hours from the time that they are added to an Amazon Kinesis stream. To add data to a stream The following example add data to the specified stream. POST HTTP1.1 Host: kinesis.. x-amz-Date: Authorization: AWS4-HMAC-SHA256 Credential=, SignedHeaders=content-type;date;host;user-agent;x-amz-date;x-amz-target;x-amzn-requestid, Signature= User-Agent: Content-Type: application/x-amz-json-1.1 Content-Length: Connection: Keep-Alive]]> X-Amz-Target: Kinesis_20131202.PutRecord { StreamName: "exampleStreamName", Data: XzxkYXRhPl8x, PartitionKey: "partitionKey" } HTTP/1.1 200 OK x-amzn-RequestId: Content-Type: application/x-amz-json-1.1 Content-Length: Date: ]]> { SequenceNumber: "21269319989653637946712965403778482177", ShardId: "shardId-000000000001" }.

Request

data PutRecord

Represents the input for PutRecord.

Request constructor

putRecord

Arguments

:: Text
prStreamName

-> Base64
prData

-> Text
prPartitionKey

-> PutRecord

Smart constructor for the minimum required parameters to construct a valid PutRecord request.

The fields accessible through corresponding lenses are:

StreamName :: Text

Data :: Base64

PartitionKey :: Text

ExplicitHashKey :: Maybe Text

SequenceNumberForOrdering :: Maybe Text

Request lenses

prStreamName :: Lens' PutRecord Text

The name of the stream to put the data record into.

prData :: Lens' PutRecord Base64

The data blob to put into the record, which is Base64-encoded when the blob is serialized. The maximum size of the data blob (the payload after Base64-decoding) is 50 kilobytes (KB).

prPartitionKey :: Lens' PutRecord Text

Determines which shard in the stream the data record is assigned to. Partition keys are Unicode strings with a maximum length limit of 256 bytes. Amazon Kinesis uses the partition key as input to a hash function that maps the partition key and associated data to a specific shard. Specifically, an MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key will map to the same shard within the stream.

prExplicitHashKey :: Lens' PutRecord (Maybe Text)

The hash value used to explicitly determine the shard the data record is assigned to by overriding the partition key hash.

prSequenceNumberForOrdering :: Lens' PutRecord (Maybe Text)

Guarantees strictly increasing sequence numbers, for puts from the same client and to the same partition key. Usage: set the SequenceNumberForOrdering of record n to the sequence number of record n-1 (as returned in the PutRecordResult when putting record n-1). If this parameter is not set, records will be coarsely ordered based on arrival time.

Response

data PutRecordResponse

Represents the output for PutRecord.

Response constructor

putRecordResponse

Arguments

:: Text
prrShardId

-> Text
prrSequenceNumber

-> PutRecordResponse

Smart constructor for the minimum required parameters to construct a valid PutRecordResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ShardId :: Text

SequenceNumber :: Text

Response lenses

prrShardId :: Lens' PutRecordResponse Text

The shard ID of the shard where the data record was placed.

prrSequenceNumber :: Lens' PutRecordResponse Text

The sequence number identifier that was assigned to the put data record. The sequence number for the record is unique across all records in the stream. A sequence number is the identifier associated with every record put into the stream.