Network.AWS.Kinesis.RemoveTagsFromStream

Description

Removes tags from the specified Amazon Kinesis stream. Removed tags are deleted and cannot be recovered after this operation successfully completes.

If you specify a tag that does not exist, it is ignored.

Creating a Request

removeTagsFromStream

Arguments

:: Text
rtfsStreamName

-> NonEmpty Text
rtfsTagKeys

-> RemoveTagsFromStream

Creates a value of RemoveTagsFromStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtfsStreamName - The name of the stream.

rtfsTagKeys - A list of tag keys. Each corresponding tag is removed from the stream.

data RemoveTagsFromStream

Represents the input for RemoveTagsFromStream .

See: removeTagsFromStream smart constructor.

Request Lenses

rtfsStreamName :: Lens' RemoveTagsFromStream Text

The name of the stream.

rtfsTagKeys :: Lens' RemoveTagsFromStream (NonEmpty Text)

A list of tag keys. Each corresponding tag is removed from the stream.

Destructuring the Response

removeTagsFromStreamResponse :: RemoveTagsFromStreamResponse

Creates a value of RemoveTagsFromStreamResponse with the minimum fields required to make a request.

data RemoveTagsFromStreamResponse

See: removeTagsFromStreamResponse smart constructor.