Network.AWS.Kinesis.StopStreamEncryption

Description

Disables server-side encryption for a specified stream.

Stopping encryption is an asynchronous operation. Upon receiving the request, Amazon Kinesis returns immediately and sets the status of the stream to UPDATING . After the update is complete, Amazon Kinesis sets the status of the stream back to ACTIVE . Stopping encryption normally takes a few seconds to complete but it can take minutes. You can continue to read and write data to your stream while its status is UPDATING . Once the status of the stream is ACTIVE records written to the stream will no longer be encrypted by the Amazon Kinesis Streams service.

API Limits: You can successfully disable server-side encryption 25 times in a rolling 24 hour period.

Note: It can take up to 5 seconds after the stream is in an ACTIVE status before all records written to the stream are no longer subject to encryption. After you’ve disabled encryption, you can verify encryption was not applied by inspecting the API response from PutRecord or PutRecords .

Creating a Request

stopStreamEncryption

Arguments

:: Text
sseStreamName

-> EncryptionType
sseEncryptionType

-> Text
sseKeyId

-> StopStreamEncryption

Creates a value of StopStreamEncryption with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sseStreamName - The name of the stream on which to stop encrypting records.

sseEncryptionType - The encryption type. This parameter can be one of the following values: * NONE : Not valid for this operation. An InvalidOperationException will be thrown. * KMS : Use server-side encryption on the records in the stream using a customer-managed KMS key.

sseKeyId - The GUID for the customer-managed key that was used for encryption.

data StopStreamEncryption

See: stopStreamEncryption smart constructor.

Request Lenses

sseStreamName :: Lens' StopStreamEncryption Text

The name of the stream on which to stop encrypting records.

sseEncryptionType :: Lens' StopStreamEncryption EncryptionType

The encryption type. This parameter can be one of the following values: * NONE : Not valid for this operation. An InvalidOperationException will be thrown. * KMS : Use server-side encryption on the records in the stream using a customer-managed KMS key.

sseKeyId :: Lens' StopStreamEncryption Text

The GUID for the customer-managed key that was used for encryption.

Destructuring the Response

stopStreamEncryptionResponse :: StopStreamEncryptionResponse

Creates a value of StopStreamEncryptionResponse with the minimum fields required to make a request.

data StopStreamEncryptionResponse

See: stopStreamEncryptionResponse smart constructor.