Network.AWS.Kinesis.Types

Description

Amazon Kinesis is a managed service that scales elastically for real-time processing of streaming big data. The service takes in large streams of data records that can then be consumed in real time by multiple data-processing applications that can be run on Amazon EC2 instances.

Service

data Kinesis

Supported version (2013-12-02) of the Amazon Kinesis service.

Errors

data KinesisError

A sum type representing possible errors returned by the Kinesis service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

ExpiredIteratorException
The provided iterator exceeds the maximum age allowed.

Fields

_eieMessage :: Maybe Text

InvalidArgumentException
A specified parameter exceeds its restrictions, is not supported, or can't be used. For more information, see the returned message.

Fields

_iaeMessage :: Maybe Text

KinesisClient HttpException

KinesisSerializer String

KinesisService String

LimitExceededException
The requested resource exceeds the maximum number allowed, or the number of concurrent stream requests exceeds the maximum number allowed (5).

Fields

_leeMessage :: Maybe Text

ProvisionedThroughputExceededException
The request rate is too high, or the requested data is too large for the available throughput. Reduce the frequency or size of your requests. For more information, see Error Retries and Exponential Backoff in AWS in the AWS General Reference.

Fields

_pteeMessage :: Maybe Text

ResourceInUseException
The resource is not available for this operation. For example, you attempted to split a shard but the stream is not in the ACTIVE state.

Fields

_riueMessage :: Maybe Text

ResourceNotFoundException
The requested resource could not be found. It might not be specified correctly, or it might not be in the ACTIVE state.

Fields

_rnfeMessage :: Maybe Text

_ExpiredIteratorException :: Prism' KinesisError (Maybe Text)

The provided iterator exceeds the maximum age allowed.

See: ExpiredIteratorException

_InvalidArgumentException :: Prism' KinesisError (Maybe Text)

A specified parameter exceeds its restrictions, is not supported, or can't be used. For more information, see the returned message.

See: InvalidArgumentException

_KinesisClient :: Prism' KinesisError HttpException

See: KinesisClient

_KinesisSerializer :: Prism' KinesisError String

See: KinesisSerializer

_KinesisService :: Prism' KinesisError String

See: KinesisService

_LimitExceededException :: Prism' KinesisError (Maybe Text)

The requested resource exceeds the maximum number allowed, or the number of concurrent stream requests exceeds the maximum number allowed (5).

See: LimitExceededException

_ProvisionedThroughputExceededException :: Prism' KinesisError (Maybe Text)

The request rate is too high, or the requested data is too large for the available throughput. Reduce the frequency or size of your requests. For more information, see Error Retries and Exponential Backoff in AWS in the AWS General Reference.

See: ProvisionedThroughputExceededException

_ResourceInUseException :: Prism' KinesisError (Maybe Text)

The resource is not available for this operation. For example, you attempted to split a shard but the stream is not in the ACTIVE state.

See: ResourceInUseException

_ResourceNotFoundException :: Prism' KinesisError (Maybe Text)

The requested resource could not be found. It might not be specified correctly, or it might not be in the ACTIVE state.

See: ResourceNotFoundException

ShardIteratorType

data ShardIteratorType

Constructors

ShardIteratorTypeAfterSequenceNumber
AFTER_SEQUENCE_NUMBER

ShardIteratorTypeAtSequenceNumber
AT_SEQUENCE_NUMBER

ShardIteratorTypeLatest
LATEST

ShardIteratorTypeTrimHorizon
TRIM_HORIZON

StreamStatus

data StreamStatus

Constructors

StreamStatusActive
ACTIVE

StreamStatusCreating
CREATING

StreamStatusDeleting
DELETING

StreamStatusUpdating
UPDATING

HashKeyRange

data HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

hashKeyRange

Arguments

:: Text
hkrStartingHashKey

-> Text
hkrEndingHashKey

-> HashKeyRange

Smart constructor for the minimum required fields to construct a valid HashKeyRange data type to populate a request.

The fields accessible through corresponding lenses are:

StartingHashKey :: Text

EndingHashKey :: Text

hkrStartingHashKey :: Lens' HashKeyRange Text

The starting hash key of the hash key range.

hkrEndingHashKey :: Lens' HashKeyRange Text

The ending hash key of the hash key range.

Record

data Record

The unit of data of the Amazon Kinesis stream, which is composed of a sequence number, a partition key, and a data blob.

record

Arguments

:: Text
rSequenceNumber

-> Base64
rData

-> Text
rPartitionKey

-> Record

Smart constructor for the minimum required fields to construct a valid Record data type.

Record is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SequenceNumber :: Text

Data :: Base64

PartitionKey :: Text

rSequenceNumber :: Lens' Record Text

The unique identifier for the record in the Amazon Kinesis stream.

rData :: Lens' Record Base64

The data blob. The data in the blob is both opaque and immutable to the Amazon Kinesis service, which does not inspect, interpret, or change the data in the blob in any way. The maximum size of the data blob (the payload after Base64-decoding) is 50 kilobytes (KB).

rPartitionKey :: Lens' Record Text

Identifies which shard in the stream the data record is assigned to.

SequenceNumberRange

data SequenceNumberRange

The range of possible sequence numbers for the shard.

sequenceNumberRange

Arguments

:: Text
snrStartingSequenceNumber

-> SequenceNumberRange

Smart constructor for the minimum required fields to construct a valid SequenceNumberRange data type to populate a request.

The fields accessible through corresponding lenses are:

StartingSequenceNumber :: Text

EndingSequenceNumber :: Maybe Text

snrStartingSequenceNumber :: Lens' SequenceNumberRange Text

The starting sequence number for the range.

snrEndingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The ending sequence number for the range. Shards that are in the OPEN state have an ending sequence number of null.

Shard

data Shard

A uniquely identified group of data records in an Amazon Kinesis stream.

shard

Arguments

:: Text
sShardId

-> HashKeyRange
sHashKeyRange

-> SequenceNumberRange
sSequenceNumberRange

-> Shard

Smart constructor for the minimum required fields to construct a valid Shard data type.

Shard is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ShardId :: Text

ParentShardId :: Maybe Text

AdjacentParentShardId :: Maybe Text

HashKeyRange :: HashKeyRange

SequenceNumberRange :: SequenceNumberRange

sShardId :: Lens' Shard Text

The unique identifier of the shard within the Amazon Kinesis stream.

sParentShardId :: Lens' Shard (Maybe Text)

The shard Id of the shard's parent.

sAdjacentParentShardId :: Lens' Shard (Maybe Text)

The shard Id of the shard adjacent to the shard's parent.

sHashKeyRange :: Lens' Shard HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

sSequenceNumberRange :: Lens' Shard SequenceNumberRange

The range of possible sequence numbers for the shard.

StreamDescription

data StreamDescription

The current status of the stream, the stream ARN, an array of shard objects that comprise the stream, and states whether there are more shards available.

streamDescription

Arguments

:: Text
sdStreamName

-> Text
sdStreamARN

-> StreamStatus
sdStreamStatus

-> [Shard]
sdShards

-> Bool
sdHasMoreShards

-> StreamDescription

Smart constructor for the minimum required fields to construct a valid StreamDescription data type.

StreamDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

StreamName :: Text

StreamARN :: Text

StreamStatus :: StreamStatus

Shards :: [Shard]

HasMoreShards :: Bool

sdStreamName :: Lens' StreamDescription Text

The name of the stream being described.

sdStreamARN :: Lens' StreamDescription Text

The Amazon Resource Name (ARN) for the stream being described.

sdStreamStatus :: Lens' StreamDescription StreamStatus

The current status of the stream being described. The stream status is one of the following states: CREATING - The stream is being created. Amazon Kinesis immediately returns and sets StreamStatus to CREATING. DELETING - The stream is being deleted. The specified stream is in the DELETING state until Amazon Kinesis completes the deletion. ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

sdShards :: Lens' StreamDescription [Shard]

The shards that comprise the stream.

sdHasMoreShards :: Lens' StreamDescription Bool

If set to true, more shards in the stream are available to describe.

Tag

data Tag

Metadata assigned to the stream, consisting of a key-value pair.

tag

Arguments

:: Text
tKey

-> Tag

Smart constructor for the minimum required fields to construct a valid Tag data type.

Tag is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Key :: Text

Value :: Maybe Text

tKey :: Lens' Tag Text

A unique identifier for the tag. Maximum length: 128 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @.

tValue :: Lens' Tag (Maybe Text)

An optional string, typically used to describe or define the tag. Maximum length: 256 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @.