Network.AWS.Kinesis.V2013_12_02

Description

Amazon Kinesis is a managed service that scales elastically for real-time processing of streaming big data. The service takes in large streams of data records that can then be consumed in real time by multiple data-processing applications that can be run on Amazon EC2 instances.

Documentation

data CreateStream

Constructors

CreateStream

Fields

_csiShardCount :: Integer

The number of shards that the stream will use. The throughput of the stream is a function of the number of shards; more shards are required for greater provisioned throughput. Note: The default limit for an AWS account is 10 shards per stream. If you need to create a stream with more than 10 shards, contact AWS Support to increase the limit on your account.

_csiStreamName :: Text

A name to identify the stream. The stream name is scoped to the AWS account used by the application that creates the stream. It is also scoped by region. That is, two streams in two different AWS accounts can have the same name, and two streams in the same AWS account, but in two different regions, can have the same name.

csiStreamName :: Lens' CreateStream Text

csiShardCount :: Lens' CreateStream Integer

data CreateStreamResponse

Constructors

CreateStreamResponse

data DeleteStream

Constructors

DeleteStream

Fields

_dsiStreamName :: Text

The name of the stream to delete.

dsiStreamName :: Iso' DeleteStream Text

data DeleteStreamResponse

Constructors

DeleteStreamResponse

describeStream

Arguments

:: Text
_dsjStreamName

-> DescribeStream

Minimum specification for a DescribeStream request.

data DescribeStream

Constructors

DescribeStream

Fields

_dsjStreamName :: Text

The name of the stream to describe.

_dsjLimit :: Maybe Integer

The maximum number of shards to return.

_dsjExclusiveStartShardId :: Maybe Text

The shard ID of the shard to start with for the stream description.

dsjStreamName :: Lens' DescribeStream Text

dsjLimit :: Lens' DescribeStream (Maybe Integer)

dsjExclusiveStartShardId :: Lens' DescribeStream (Maybe Text)

data DescribeStreamResponse

Constructors

DescribeStreamResponse

Fields

_dsoStreamDescription :: StreamDescription

Contains the current status of the stream, the stream ARN, an array of shard objects that comprise the stream, and states whether there are more shards available.

dsoStreamDescription :: Iso' DescribeStreamResponse StreamDescription

getRecords

Arguments

:: Text
_griShardIterator

-> GetRecords

Minimum specification for a GetRecords request.

data GetRecords

Constructors

GetRecords

Fields

_griShardIterator :: Text

The position in the shard from which you want to start sequentially reading data records.

_griLimit :: Maybe Integer

The maximum number of records to return, which can be set to a value of up to 10,000.

griShardIterator :: Lens' GetRecords Text

griLimit :: Lens' GetRecords (Maybe Integer)

data GetRecordsResponse

Constructors

GetRecordsResponse

Fields

_groRecords :: [Record]

The data records retrieved from the shard.

_groNextShardIterator :: Maybe Text

The next position in the shard from which to start sequentially reading data records. If set to null, the shard has been closed and the requested iterator will not return any more data.

groRecords :: Lens' GetRecordsResponse [Record]

groNextShardIterator :: Lens' GetRecordsResponse (Maybe Text)

getShardIterator

Arguments

:: Text
_gsiiShardId

-> ShardIteratorType
_gsiiShardIteratorType

-> Text
_gsiiStreamName

-> GetShardIterator

Minimum specification for a GetShardIterator request.

data GetShardIterator

Constructors

GetShardIterator

Fields

_gsiiShardId :: Text

The shard ID of the shard to get the iterator for.

_gsiiShardIteratorType :: ShardIteratorType

Determines how the shard iterator is used to start reading data records from the shard. The following are the valid shard iterator types: AT_SEQUENCE_NUMBER - Start reading exactly from the position denoted by a specific sequence number. AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number. TRIM_HORIZON - Start reading at the last untrimmed record in the shard in the system, which is the oldest data record in the shard. LATEST - Start reading just after the most recent record in the shard, so that you always read the most recent data in the shard.

_gsiiStreamName :: Text

The name of the stream.

_gsiiStartingSequenceNumber :: Maybe Text

The sequence number of the data record in the shard from which to start reading from.

gsiiStreamName :: Lens' GetShardIterator Text

gsiiStartingSequenceNumber :: Lens' GetShardIterator (Maybe Text)

gsiiShardIteratorType :: Lens' GetShardIterator ShardIteratorType

gsiiShardId :: Lens' GetShardIterator Text

data GetShardIteratorResponse

Constructors

GetShardIteratorResponse

Fields

_gsioShardIterator :: Maybe Text

The position in the shard from which to start reading data records sequentially. A shard iterator specifies this position using the sequence number of a data record in a shard.

gsioShardIterator :: Iso' GetShardIteratorResponse (Maybe Text)

listStreams :: ListStreams

Minimum specification for a ListStreams request.

data ListStreams

Constructors

ListStreams

Fields

_lsiLimit :: Maybe Integer

The maximum number of streams to list.

_lsiExclusiveStartStreamName :: Maybe Text

The name of the stream to start the list with.

lsiLimit :: Lens' ListStreams (Maybe Integer)

lsiExclusiveStartStreamName :: Lens' ListStreams (Maybe Text)

data ListStreamsResponse

Constructors

ListStreamsResponse

Fields

_lsoHasMoreStreams :: Bool

If set to true, there are more streams available to list.

_lsoStreamNames :: [Text]

The names of the streams that are associated with the AWS account making the ListStreams request.

lsoStreamNames :: Lens' ListStreamsResponse [Text]

lsoHasMoreStreams :: Lens' ListStreamsResponse Bool

data MergeShards

Constructors

MergeShards

Fields

_msiShardToMerge :: Text

The shard ID of the shard to combine with the adjacent shard for the merge.

_msiAdjacentShardToMerge :: Text

The shard ID of the adjacent shard for the merge.

_msiStreamName :: Text

The name of the stream for the merge.

msiStreamName :: Lens' MergeShards Text

msiShardToMerge :: Lens' MergeShards Text

msiAdjacentShardToMerge :: Lens' MergeShards Text

data MergeShardsResponse

Constructors

MergeShardsResponse

putRecord

Arguments

:: Base64
_priData

-> Text
_priPartitionKey

-> Text
_priStreamName

-> PutRecord

Minimum specification for a PutRecord request.

data PutRecord

Constructors

PutRecord

Fields

_priData :: Base64

The data blob to put into the record, which is Base64-encoded when the blob is serialized. The maximum size of the data blob (the payload after Base64-decoding) is 50 kilobytes (KB).

_priPartitionKey :: Text

Determines which shard in the stream the data record is assigned to. Partition keys are Unicode strings with a maximum length limit of 256 bytes. Amazon Kinesis uses the partition key as input to a hash function that maps the partition key and associated data to a specific shard. Specifically, an MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key will map to the same shard within the stream.

_priStreamName :: Text

The name of the stream to put the data record into.

_priExplicitHashKey :: Maybe Text

The hash value used to explicitly determine the shard the data record is assigned to by overriding the partition key hash.

_priSequenceNumberForOrdering :: Maybe Text

Guarantees strictly increasing sequence numbers, for puts from the same client and to the same partition key. Usage: set the SequenceNumberForOrdering of record n to the sequence number of record n-1 (as returned in the PutRecordResult when putting record n-1). If this parameter is not set, records will be coarsely ordered based on arrival time.

priStreamName :: Lens' PutRecord Text

priSequenceNumberForOrdering :: Lens' PutRecord (Maybe Text)

priPartitionKey :: Lens' PutRecord Text

priExplicitHashKey :: Lens' PutRecord (Maybe Text)

priData :: Lens' PutRecord Base64

data PutRecordResponse

Constructors

PutRecordResponse

Fields

_proSequenceNumber :: Text

The sequence number identifier that was assigned to the put data record. The sequence number for the record is unique across all records in the stream. A sequence number is the identifier associated with every record put into the stream.

_proShardId :: Text

The shard ID of the shard where the data record was placed.

proShardId :: Lens' PutRecordResponse Text

proSequenceNumber :: Lens' PutRecordResponse Text

data SplitShard

Constructors

SplitShard

Fields

_ssiNewStartingHashKey :: Text

A hash key value for the starting hash key of one of the child shards created by the split. The hash key range for a given shard constitutes a set of ordered contiguous positive integers. The value for NewStartingHashKey must be in the range of hash keys being mapped into the shard. The NewStartingHashKey hash key value and all higher hash key values in hash key range are distributed to one of the child shards. All the lower hash key values in the range are distributed to the other child shard.

_ssiShardToSplit :: Text

The shard ID of the shard to split.

_ssiStreamName :: Text

The name of the stream for the shard split.

ssiStreamName :: Lens' SplitShard Text

ssiShardToSplit :: Lens' SplitShard Text

ssiNewStartingHashKey :: Lens' SplitShard Text

data SplitShardResponse

Constructors

SplitShardResponse

data Kinesis

Supported version (2013-12-02) of the Amazon Kinesis service.

data ShardIteratorType

Determines how the shard iterator is used to start reading data records from the shard. The following are the valid shard iterator types: AT_SEQUENCE_NUMBER - Start reading exactly from the position denoted by a specific sequence number. AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number. TRIM_HORIZON - Start reading at the last untrimmed record in the shard in the system, which is the oldest data record in the shard. LATEST - Start reading just after the most recent record in the shard, so that you always read the most recent data in the shard.

Constructors

ShardIteratorTypeAfterSequenceNumber
AFTER_SEQUENCE_NUMBER

ShardIteratorTypeAtSequenceNumber
AT_SEQUENCE_NUMBER

ShardIteratorTypeLatest
LATEST

ShardIteratorTypeTrimHorizon
TRIM_HORIZON

data StreamStatus

The current status of the stream being described. The stream status is one of the following states: CREATING - The stream is being created. Upon receiving a CreateStream request, Amazon Kinesis immediately returns and sets StreamStatus to CREATING. DELETING - The stream is being deleted. After a DeleteStream request, the specified stream is in the DELETING state until Amazon Kinesis completes the deletion. ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

Constructors

StreamStatusActive
ACTIVE

StreamStatusCreating
CREATING

StreamStatusDeleting
DELETING

StreamStatusUpdating
UPDATING

data HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

Constructors

HashKeyRange

Fields

_hkrEndingHashKey :: Text

The ending hash key of the hash key range.

_hkrStartingHashKey :: Text

The starting hash key of the hash key range.

data Record

The unit of data of the Amazon Kinesis stream, which is composed of a sequence number, a partition key, and a data blob.

Constructors

Record

Fields

_rSequenceNumber :: Text

The unique identifier for the record in the Amazon Kinesis stream.

_rPartitionKey :: Text

Identifies which shard in the stream the data record is assigned to.

_rData :: Base64

The data blob. The data in the blob is both opaque and immutable to the Amazon Kinesis service, which does not inspect, interpret, or change the data in the blob in any way. The maximum size of the data blob (the payload after Base64-decoding) is 50 kilobytes (KB).

data SequenceNumberRange

The range of possible sequence numbers for the shard.

Constructors

SequenceNumberRange

Fields

_snrStartingSequenceNumber :: Text

The starting sequence number for the range.

_snrEndingSequenceNumber :: Maybe Text

The ending sequence number for the range. Shards that are in the OPEN state have an ending sequence number of null.

data Shard

A uniquely identified group of data records in an Amazon Kinesis stream.

Constructors

Shard

Fields

_sAdjacentParentShardId :: Maybe Text

The shard Id of the shard adjacent to the shard's parent.

_sHashKeyRange :: HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

_sParentShardId :: Maybe Text

The shard Id of the shard's parent.

_sSequenceNumberRange :: SequenceNumberRange

The range of possible sequence numbers for the shard.

_sShardId :: Text

The unique identifier of the shard within the Amazon Kinesis stream.

data StreamDescription

Contains the current status of the stream, the stream ARN, an array of shard objects that comprise the stream, and states whether there are more shards available.

Constructors

StreamDescription

Fields

_sdStreamStatus :: StreamStatus

The current status of the stream being described. The stream status is one of the following states: CREATING - The stream is being created. Upon receiving a CreateStream request, Amazon Kinesis immediately returns and sets StreamStatus to CREATING. DELETING - The stream is being deleted. After a DeleteStream request, the specified stream is in the DELETING state until Amazon Kinesis completes the deletion. ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

_sdHasMoreShards :: Bool

If set to true there are more shards in the stream available to describe.

_sdStreamARN :: Text

The Amazon Resource Name (ARN) for the stream being described.

_sdShards :: [Shard]

The shards that comprise the stream.

_sdStreamName :: Text

The name of the stream being described.

hkrStartingHashKey :: Lens' HashKeyRange Text

hkrEndingHashKey :: Lens' HashKeyRange Text

rSequenceNumber :: Lens' Record Text

rPartitionKey :: Lens' Record Text

rData :: Lens' Record Base64

snrStartingSequenceNumber :: Lens' SequenceNumberRange Text

snrEndingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

sShardId :: Lens' Shard Text

sSequenceNumberRange :: Lens' Shard SequenceNumberRange

sParentShardId :: Lens' Shard (Maybe Text)

sHashKeyRange :: Lens' Shard HashKeyRange

sAdjacentParentShardId :: Lens' Shard (Maybe Text)

sdStreamStatus :: Lens' StreamDescription StreamStatus

sdStreamName :: Lens' StreamDescription Text

sdStreamARN :: Lens' StreamDescription Text

sdShards :: Lens' StreamDescription [Shard]

sdHasMoreShards :: Lens' StreamDescription Bool