Network.AWS.Kinesis.CreateStream

Description

Creates an Amazon Kinesis stream. A stream captures and transports data records that are continuously emitted from different data sources or producers . Scale-out within a stream is explicitly supported by means of shards, which are uniquely identified groups of data records in a stream.

You specify and control the number of shards that a stream is composed of. Each shard can support reads up to 5 transactions per second, up to a maximum data read total of 2 MB per second. Each shard can support writes up to 1,000 records per second, up to a maximum data write total of 1 MB per second. You can add shards to a stream if the amount of data input increases and you can remove shards if the amount of data input decreases.

The stream name identifies the stream. The name is scoped to the AWS account used by the application. It is also scoped by region. That is, two streams in two different accounts can have the same name, and two streams in the same account, but in two different regions, can have the same name.

CreateStream is an asynchronous operation. Upon receiving a CreateStream request, Amazon Kinesis immediately returns and sets the stream status to CREATING . After the stream is created, Amazon Kinesis sets the stream status to ACTIVE . You should perform read and write operations only on an ACTIVE stream.

You receive a LimitExceededException when making a CreateStream request if you try to do one of the following:

Have more than five streams in the CREATING state at any point in time.

Create more shards than are authorized for your account.

For the default shard limit for an AWS account, see Streams Limits in the Amazon Kinesis Streams Developer Guide . If you need to increase this limit, contact AWS Support .

You can use DescribeStream to check the stream status, which is returned in StreamStatus .

CreateStream has a limit of 5 transactions per second per account.

Creating a Request

createStream

Arguments

:: Text
csStreamName

-> Natural
csShardCount

-> CreateStream

Creates a value of CreateStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csStreamName - A name to identify the stream. The stream name is scoped to the AWS account used by the application that creates the stream. It is also scoped by region. That is, two streams in two different AWS accounts can have the same name, and two streams in the same AWS account but in two different regions can have the same name.

csShardCount - The number of shards that the stream will use. The throughput of the stream is a function of the number of shards; more shards are required for greater provisioned throughput. DefaultShardLimit;

data CreateStream

Represents the input for CreateStream .

See: createStream smart constructor.

Request Lenses

csStreamName :: Lens' CreateStream Text

A name to identify the stream. The stream name is scoped to the AWS account used by the application that creates the stream. It is also scoped by region. That is, two streams in two different AWS accounts can have the same name, and two streams in the same AWS account but in two different regions can have the same name.

csShardCount :: Lens' CreateStream Natural

The number of shards that the stream will use. The throughput of the stream is a function of the number of shards; more shards are required for greater provisioned throughput. DefaultShardLimit;

Destructuring the Response

createStreamResponse :: CreateStreamResponse

Creates a value of CreateStreamResponse with the minimum fields required to make a request.

data CreateStreamResponse

See: createStreamResponse smart constructor.