Network.AWS.Kinesis.DeleteStream

Description

Deletes a stream and all its shards and data. You must shut down any applications that are operating on the stream before you delete the stream. If an application attempts to operate on a deleted stream, it will receive the exception ResourceNotFoundException.

If the stream is in the ACTIVE state, you can delete it. After a DeleteStream request, the specified stream is in the DELETING state until Amazon Kinesis completes the deletion.

Note: Amazon Kinesis might continue to accept data read and write operations, such as PutRecord, PutRecords, and GetRecords, on a stream in the DELETING state until the stream deletion is complete.

When you delete a stream, any shards in that stream are also deleted, and any tags are dissociated from the stream.

You can use the DescribeStream operation to check the state of the stream, which is returned in StreamStatus.

DeleteStream has a limit of 5 transactions per second per account.

http://docs.aws.amazon.com/kinesis/latest/APIReference/API_DeleteStream.html

Request

data DeleteStream

Request constructor

deleteStream

Arguments

:: Text
dsStreamName

-> DeleteStream

DeleteStream constructor.

The fields accessible through corresponding lenses are:

dsStreamName :: Text

Request lenses

dsStreamName :: Lens' DeleteStream Text

The name of the stream to delete.

Response

data DeleteStreamResponse

Response constructor

deleteStreamResponse :: DeleteStreamResponse

DeleteStreamResponse constructor.