Network.AWS.Kinesis.DescribeStream

Description

Describes the specified Amazon Kinesis stream.

The information returned includes the stream name, Amazon Resource Name (ARN), creation time, enhanced metric configuration, and shard map. The shard map is an array of shard objects. For each shard object, there is the hash key and sequence number ranges that the shard spans, and the IDs of any earlier shards that played in a role in creating the shard. Every record ingested in the stream is identified by a sequence number, which is assigned when the record is put into the stream.

You can limit the number of shards returned by each call. For more information, see Retrieving Shards from a Stream in the Amazon Kinesis Streams Developer Guide .

There are no guarantees about the chronological order shards returned. To process shards in chronological order, use the ID of the parent shard to track the lineage to the oldest shard.

This operation has a limit of 10 transactions per second per account.

This operation returns paginated results.

Creating a Request

describeStream

Arguments

:: Text
dStreamName

-> DescribeStream

Creates a value of DescribeStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dExclusiveStartShardId - The shard ID of the shard to start with.

dLimit - The maximum number of shards to return in a single call. The default value is 100. If you specify a value greater than 100, at most 100 shards are returned.

dStreamName - The name of the stream to describe.

data DescribeStream

Represents the input for DescribeStream .

See: describeStream smart constructor.

Request Lenses

dExclusiveStartShardId :: Lens' DescribeStream (Maybe Text)

The shard ID of the shard to start with.

dLimit :: Lens' DescribeStream (Maybe Natural)

The maximum number of shards to return in a single call. The default value is 100. If you specify a value greater than 100, at most 100 shards are returned.

dStreamName :: Lens' DescribeStream Text

The name of the stream to describe.

Destructuring the Response

describeStreamResponse

Arguments

:: Int
dsrsResponseStatus

-> StreamDescription
dsrsStreamDescription

-> DescribeStreamResponse

Creates a value of DescribeStreamResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsResponseStatus - -- | The response status code.

dsrsStreamDescription - The current status of the stream, the stream ARN, an array of shard objects that comprise the stream, and whether there are more shards available.

data DescribeStreamResponse

Represents the output for DescribeStream .

See: describeStreamResponse smart constructor.

Response Lenses

dsrsResponseStatus :: Lens' DescribeStreamResponse Int

- | The response status code.

dsrsStreamDescription :: Lens' DescribeStreamResponse StreamDescription

The current status of the stream, the stream ARN, an array of shard objects that comprise the stream, and whether there are more shards available.