Network.AWS.Kinesis.DescribeStream

Description

Describes the specified stream.

The information about the stream includes its current status, its Amazon Resource Name (ARN), and an array of shard objects. For each shard object, there is information about the hash key and sequence number ranges that the shard spans, and the IDs of any earlier shards that played in a role in creating the shard. A sequence number is the identifier associated with every record ingested in the Amazon Kinesis stream. The sequence number is assigned when a record is put into the stream.

You can limit the number of returned shards using the Limit parameter. The number of shards in a stream may be too large to return from a single call to DescribeStream. You can detect this by using the HasMoreShards flag in the returned output. HasMoreShards is set to true when there is more data available.

DescribeStream is a paginated operation. If there are more shards available, you can request them using the shard ID of the last shard returned. Specify this ID in the ExclusiveStartShardId parameter in a subsequent request to DescribeStream.

DescribeStream has a limit of 10 transactions per second per account.

http://docs.aws.amazon.com/kinesis/latest/APIReference/API_DescribeStream.html

Request

data DescribeStream

Request constructor

describeStream

Arguments

:: Text
ds1StreamName

-> DescribeStream

DescribeStream constructor.

The fields accessible through corresponding lenses are:

ds1ExclusiveStartShardId :: Maybe Text

ds1Limit :: Maybe Natural

ds1StreamName :: Text

Request lenses

ds1ExclusiveStartShardId :: Lens' DescribeStream (Maybe Text)

The shard ID of the shard to start with.

ds1Limit :: Lens' DescribeStream (Maybe Natural)

The maximum number of shards to return.

ds1StreamName :: Lens' DescribeStream Text

The name of the stream to describe.

Response

data DescribeStreamResponse

Response constructor

describeStreamResponse

Arguments

:: StreamDescription
dsrStreamDescription

-> DescribeStreamResponse

DescribeStreamResponse constructor.

The fields accessible through corresponding lenses are:

dsrStreamDescription :: StreamDescription

Response lenses

dsrStreamDescription :: Lens' DescribeStreamResponse StreamDescription

The current status of the stream, the stream ARN, an array of shard objects that comprise the stream, and states whether there are more shards available.