Network.AWS.Kinesis.ListStreams

Description

Lists your streams.

The number of streams may be too large to return from a single call to ListStreams. You can limit the number of returned streams using the Limit parameter. If you do not specify a value for the Limit parameter, Amazon Kinesis uses the default limit, which is currently 10.

You can detect if there are more streams available to list by using the HasMoreStreams flag from the returned output. If there are more streams available, you can request more streams by using the name of the last stream returned by the ListStreams request in the ExclusiveStartStreamName parameter in a subsequent request to ListStreams. The group of stream names returned by the subsequent request is then added to the list. You can continue this process until all the stream names have been collected in the list.

ListStreams has a limit of 5 transactions per second per account.

http://docs.aws.amazon.com/kinesis/latest/APIReference/API_ListStreams.html

Request

data ListStreams

Request constructor

listStreams :: ListStreams

ListStreams constructor.

The fields accessible through corresponding lenses are:

lsExclusiveStartStreamName :: Maybe Text

lsLimit :: Maybe Natural

Request lenses

lsExclusiveStartStreamName :: Lens' ListStreams (Maybe Text)

The name of the stream to start the list with.

lsLimit :: Lens' ListStreams (Maybe Natural)

The maximum number of streams to list.

Response

data ListStreamsResponse

Response constructor

listStreamsResponse

Arguments

:: Bool
lsrHasMoreStreams

-> ListStreamsResponse

ListStreamsResponse constructor.

The fields accessible through corresponding lenses are:

lsrHasMoreStreams :: Bool

lsrStreamNames :: [Text]

Response lenses

lsrHasMoreStreams :: Lens' ListStreamsResponse Bool

If set to true, there are more streams available to list.

lsrStreamNames :: Lens' ListStreamsResponse [Text]

The names of the streams that are associated with the AWS account making the ListStreams request.