Network.AWS.Kinesis.ListTagsForStream

Description

Lists the tags for the specified Amazon Kinesis stream.

Creating a Request

listTagsForStream

Arguments

:: Text
ltfsStreamName

-> ListTagsForStream

Creates a value of ListTagsForStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltfsLimit - The number of tags to return. If this number is less than the total number of tags associated with the stream, HasMoreTags is set to true . To list additional tags, set ExclusiveStartTagKey to the last key in the response.

ltfsExclusiveStartTagKey - The key to use as the starting point for the list of tags. If this parameter is set, ListTagsForStream gets all tags that occur after ExclusiveStartTagKey .

ltfsStreamName - The name of the stream.

data ListTagsForStream

Represents the input for ListTagsForStream .

See: listTagsForStream smart constructor.

Request Lenses

ltfsLimit :: Lens' ListTagsForStream (Maybe Natural)

The number of tags to return. If this number is less than the total number of tags associated with the stream, HasMoreTags is set to true . To list additional tags, set ExclusiveStartTagKey to the last key in the response.

ltfsExclusiveStartTagKey :: Lens' ListTagsForStream (Maybe Text)

The key to use as the starting point for the list of tags. If this parameter is set, ListTagsForStream gets all tags that occur after ExclusiveStartTagKey .

ltfsStreamName :: Lens' ListTagsForStream Text

The name of the stream.

Destructuring the Response

listTagsForStreamResponse

Arguments

:: Int
ltfsrsResponseStatus

-> Bool
ltfsrsHasMoreTags

-> ListTagsForStreamResponse

Creates a value of ListTagsForStreamResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltfsrsResponseStatus - -- | The response status code.

ltfsrsTags - A list of tags associated with StreamName , starting with the first tag after ExclusiveStartTagKey and up to the specified Limit .

ltfsrsHasMoreTags - If set to true , more tags are available. To request additional tags, set ExclusiveStartTagKey to the key of the last tag returned.

data ListTagsForStreamResponse

Represents the output for ListTagsForStream .

See: listTagsForStreamResponse smart constructor.

Response Lenses

ltfsrsResponseStatus :: Lens' ListTagsForStreamResponse Int

- | The response status code.

ltfsrsTags :: Lens' ListTagsForStreamResponse [Tag]

A list of tags associated with StreamName , starting with the first tag after ExclusiveStartTagKey and up to the specified Limit .

ltfsrsHasMoreTags :: Lens' ListTagsForStreamResponse Bool

If set to true , more tags are available. To request additional tags, set ExclusiveStartTagKey to the key of the last tag returned.