Network.AWS.Kinesis.PutRecord

Description

Puts (writes) a single data record from a producer into an Amazon Kinesis stream. Call PutRecord to send data from the producer into the Amazon Kinesis stream for real-time ingestion and subsequent processing, one record at a time. Each shard can support up to 1000 records written per second, up to a maximum total of 1 MB data written per second.

You must specify the name of the stream that captures, stores, and transports the data; a partition key; and the data blob itself.

The data blob can be any type of data; for example, a segment from a log file, geographic/location data, website clickstream data, and so on.

The partition key is used by Amazon Kinesis to distribute data across shards. Amazon Kinesis segregates the data records that belong to a data stream into multiple shards, using the partition key associated with each data record to determine which shard a given data record belongs to.

Partition keys are Unicode strings, with a maximum length limit of 256 bytes. An MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards using the hash key ranges of the shards. You can override hashing the partition key to determine the shard by explicitly specifying a hash value using the ExplicitHashKey parameter. For more information, see Partition Key in the Amazon KinesisDeveloper Guide.

PutRecord returns the shard ID of where the data record was placed and the sequence number that was assigned to the data record.

Sequence numbers generally increase over time. To guarantee strictly increasing ordering, use the SequenceNumberForOrdering parameter. For more information, see Sequence Number in the Amazon Kinesis Developer Guide.

If a PutRecord request cannot be processed because of insufficient provisioned throughput on the shard involved in the request, PutRecord throws ProvisionedThroughputExceededException.

Data records are accessible for only 24 hours from the time that they are added to an Amazon Kinesis stream.

http://docs.aws.amazon.com/kinesis/latest/APIReference/API_PutRecord.html

Request

data PutRecord

Request constructor

putRecord

Arguments

:: Text
prStreamName

-> Base64
prData

-> Text
prPartitionKey

-> PutRecord

PutRecord constructor.

The fields accessible through corresponding lenses are:

prData :: Base64

prExplicitHashKey :: Maybe Text

prPartitionKey :: Text

prSequenceNumberForOrdering :: Maybe Text

prStreamName :: Text

Request lenses

prData :: Lens' PutRecord Base64

The data blob to put into the record, which is base64-encoded when the blob is serialized. The maximum size of the data blob (the payload before base64-encoding) is 50 kilobytes (KB)

prExplicitHashKey :: Lens' PutRecord (Maybe Text)

The hash value used to explicitly determine the shard the data record is assigned to by overriding the partition key hash.

prPartitionKey :: Lens' PutRecord Text

Determines which shard in the stream the data record is assigned to. Partition keys are Unicode strings with a maximum length limit of 256 bytes. Amazon Kinesis uses the partition key as input to a hash function that maps the partition key and associated data to a specific shard. Specifically, an MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key will map to the same shard within the stream.

prSequenceNumberForOrdering :: Lens' PutRecord (Maybe Text)

Guarantees strictly increasing sequence numbers, for puts from the same client and to the same partition key. Usage: set the SequenceNumberForOrdering of record n to the sequence number of record n-1 (as returned in the PutRecordResult when putting record n-1). If this parameter is not set, records will be coarsely ordered based on arrival time.

prStreamName :: Lens' PutRecord Text

The name of the stream to put the data record into.

Response

data PutRecordResponse

Response constructor

putRecordResponse

Arguments

:: Text
prrShardId

-> Text
prrSequenceNumber

-> PutRecordResponse

PutRecordResponse constructor.

The fields accessible through corresponding lenses are:

prrSequenceNumber :: Text

prrShardId :: Text

Response lenses

prrSequenceNumber :: Lens' PutRecordResponse Text

The sequence number identifier that was assigned to the put data record. The sequence number for the record is unique across all records in the stream. A sequence number is the identifier associated with every record put into the stream.

prrShardId :: Lens' PutRecordResponse Text

The shard ID of the shard where the data record was placed.