Network.AWS.Kinesis.PutRecords

Description

Writes multiple data records into an Amazon Kinesis stream in a single call (also referred to as a PutRecords request). Use this operation to send data into the stream for data ingestion and processing.

Each PutRecords request can support up to 500 records. Each record in the request can be as large as 1 MB, up to a limit of 5 MB for the entire request, including partition keys. Each shard can support writes up to 1,000 records per second, up to a maximum data write total of 1 MB per second.

You must specify the name of the stream that captures, stores, and transports the data; and an array of request Records , with each record in the array requiring a partition key and data blob. The record size limit applies to the total size of the partition key and data blob.

The data blob can be any type of data; for example, a segment from a log file, geographic/location data, website clickstream data, and so on.

The partition key is used by Amazon Kinesis as input to a hash function that maps the partition key and associated data to a specific shard. An MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key map to the same shard within the stream. For more information, see Adding Data to a Stream in the Amazon Kinesis Streams Developer Guide .

Each record in the Records array may include an optional parameter, ExplicitHashKey , which overrides the partition key to shard mapping. This parameter allows a data producer to determine explicitly the shard where the record is stored. For more information, see Adding Multiple Records with PutRecords in the Amazon Kinesis Streams Developer Guide .

The PutRecords response includes an array of response Records . Each record in the response array directly correlates with a record in the request array using natural ordering, from the top to the bottom of the request and response. The response Records array always includes the same number of records as the request array.

The response Records array includes both successfully and unsuccessfully processed records. Amazon Kinesis attempts to process all records in each PutRecords request. A single record failure does not stop the processing of subsequent records.

A successfully-processed record includes ShardId and SequenceNumber values. The ShardId parameter identifies the shard in the stream where the record is stored. The SequenceNumber parameter is an identifier assigned to the put record, unique to all records in the stream.

An unsuccessfully-processed record includes ErrorCode and ErrorMessage values. ErrorCode reflects the type of error and can be one of the following values: ProvisionedThroughputExceededException or InternalFailure . ErrorMessage provides more detailed information about the ProvisionedThroughputExceededException exception including the account ID, stream name, and shard ID of the record that was throttled. For more information about partially successful responses, see Adding Multiple Records with PutRecords in the Amazon Kinesis Streams Developer Guide .

By default, data records are accessible for 24 hours from the time that they are added to a stream. You can use IncreaseStreamRetentionPeriod or DecreaseStreamRetentionPeriod to modify this retention period.

Creating a Request

putRecords

Arguments

:: NonEmpty PutRecordsRequestEntry
pRecordEntries

-> Text
pStreamName

-> PutRecords

Creates a value of PutRecords with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pRecordEntries - The records associated with the request.

pStreamName - The stream name associated with the request.

data PutRecords

A PutRecords request.

See: putRecords smart constructor.

Request Lenses

pRecordEntries :: Lens' PutRecords (NonEmpty PutRecordsRequestEntry)

The records associated with the request.

pStreamName :: Lens' PutRecords Text

The stream name associated with the request.

Destructuring the Response

putRecordsResponse

Arguments

:: Int
prsResponseStatus

-> NonEmpty PutRecordsResultEntry
prsRecords

-> PutRecordsResponse

Creates a value of PutRecordsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prsEncryptionType - The encryption type used on the records. This parameter can be one of the following values: * NONE : Do not encrypt the records. * KMS : Use server-side encryption on the records using a customer-managed KMS key.

prsFailedRecordCount - The number of unsuccessfully processed records in a PutRecords request.

prsResponseStatus - -- | The response status code.

prsRecords - An array of successfully and unsuccessfully processed record results, correlated with the request by natural ordering. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to a stream includes ErrorCode and ErrorMessage in the result.

data PutRecordsResponse

PutRecords results.

See: putRecordsResponse smart constructor.

Response Lenses

prsEncryptionType :: Lens' PutRecordsResponse (Maybe EncryptionType)

The encryption type used on the records. This parameter can be one of the following values: * NONE : Do not encrypt the records. * KMS : Use server-side encryption on the records using a customer-managed KMS key.

prsFailedRecordCount :: Lens' PutRecordsResponse (Maybe Natural)

The number of unsuccessfully processed records in a PutRecords request.

prsResponseStatus :: Lens' PutRecordsResponse Int

- | The response status code.

prsRecords :: Lens' PutRecordsResponse (NonEmpty PutRecordsResultEntry)

An array of successfully and unsuccessfully processed record results, correlated with the request by natural ordering. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to a stream includes ErrorCode and ErrorMessage in the result.