Network.AWS.Kinesis.SplitShard

Description

Splits a shard into two new shards in the Amazon Kinesis stream to increase the stream's capacity to ingest and transport data. SplitShard is called when there is a need to increase the overall capacity of a stream because of an expected increase in the volume of data records being ingested.

You can also use SplitShard when a shard appears to be approaching its maximum utilization; for example, the producers sending data into the specific shard are suddenly sending more than previously anticipated. You can also call SplitShard to increase stream capacity, so that more Amazon Kinesis applications can simultaneously read data from the stream for real-time processing.

You must specify the shard to be split and the new hash key, which is the position in the shard where the shard gets split in two. In many cases, the new hash key might simply be the average of the beginning and ending hash key, but it can be any hash key value in the range being mapped into the shard. For more information about splitting shards, see Split a Shard in the Amazon Kinesis Streams Developer Guide .

You can use DescribeStream to determine the shard ID and hash key values for the ShardToSplit and NewStartingHashKey parameters that are specified in the SplitShard request.

SplitShard is an asynchronous operation. Upon receiving a SplitShard request, Amazon Kinesis immediately returns a response and sets the stream status to UPDATING . After the operation is completed, Amazon Kinesis sets the stream status to ACTIVE . Read and write operations continue to work while the stream is in the UPDATING state.

You can use DescribeStream to check the status of the stream, which is returned in StreamStatus . If the stream is in the ACTIVE state, you can call SplitShard . If a stream is in CREATING or UPDATING or DELETING states, DescribeStream returns a ResourceInUseException .

If the specified stream does not exist, DescribeStream returns a ResourceNotFoundException . If you try to create more shards than are authorized for your account, you receive a LimitExceededException .

For the default shard limit for an AWS account, see Streams Limits in the Amazon Kinesis Streams Developer Guide . If you need to increase this limit, contact AWS Support .

If you try to operate on too many streams simultaneously using CreateStream , DeleteStream , MergeShards , and/or SplitShard , you receive a LimitExceededException .

SplitShard has limit of 5 transactions per second per account.

Creating a Request

splitShard

Arguments

:: Text
ssStreamName

-> Text
ssShardToSplit

-> Text
ssNewStartingHashKey

-> SplitShard

Creates a value of SplitShard with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssStreamName - The name of the stream for the shard split.

ssShardToSplit - The shard ID of the shard to split.

ssNewStartingHashKey - A hash key value for the starting hash key of one of the child shards created by the split. The hash key range for a given shard constitutes a set of ordered contiguous positive integers. The value for NewStartingHashKey must be in the range of hash keys being mapped into the shard. The NewStartingHashKey hash key value and all higher hash key values in hash key range are distributed to one of the child shards. All the lower hash key values in the range are distributed to the other child shard.

data SplitShard

Represents the input for SplitShard .

See: splitShard smart constructor.

Request Lenses

ssStreamName :: Lens' SplitShard Text

The name of the stream for the shard split.

ssShardToSplit :: Lens' SplitShard Text

The shard ID of the shard to split.

ssNewStartingHashKey :: Lens' SplitShard Text

A hash key value for the starting hash key of one of the child shards created by the split. The hash key range for a given shard constitutes a set of ordered contiguous positive integers. The value for NewStartingHashKey must be in the range of hash keys being mapped into the shard. The NewStartingHashKey hash key value and all higher hash key values in hash key range are distributed to one of the child shards. All the lower hash key values in the range are distributed to the other child shard.

Destructuring the Response

splitShardResponse :: SplitShardResponse

Creates a value of SplitShardResponse with the minimum fields required to make a request.

data SplitShardResponse

See: splitShardResponse smart constructor.