Network.AWS.Kinesis.Types

Description

Service Configuration

kinesis :: Service

API version 2013-12-02 of the Amazon Kinesis SDK configuration.

Errors

_KMSInvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the state of the specified resource isn't valid for this request. For more information, see How Key State Affects Use of a Customer Master Key in the AWS Key Management Service Developer Guide .

_KMSThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling. For more information about throttling, see Limits in the AWS Key Management Service Developer Guide .

_ExpiredIteratorException :: AsError a => Getting (First ServiceError) a ServiceError

The provided iterator exceeds the maximum age allowed.

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

A specified parameter exceeds its restrictions, is not supported, or can't be used. For more information, see the returned message.

_KMSOptInRequired :: AsError a => Getting (First ServiceError) a ServiceError

The AWS access key ID needs a subscription for the service.

_ProvisionedThroughputExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request rate for the stream is too high, or the requested data is too large for the available throughput. Reduce the frequency or size of your requests. For more information, see Streams Limits in the Amazon Kinesis Streams Developer Guide , and Error Retries and Exponential Backoff in AWS in the AWS General Reference .

_KMSNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified entity or resource couldn't be found.

_KMSDisabledException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified CMK isn't enabled.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource could not be found. The stream might not be specified correctly.

_KMSAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The ciphertext references a key that doesn't exist or that you don't have access to.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource exceeds the maximum number allowed, or the number of concurrent stream requests exceeds the maximum number allowed (5).

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is not available for this operation. For successful operation, the resource needs to be in the ACTIVE state.

EncryptionType

data EncryptionType

Constructors

KMS

None

MetricsName

data MetricsName

Constructors

All

IncomingBytes

IncomingRecords

IteratorAgeMilliseconds

OutgoingBytes

OutgoingRecords

ReadProvisionedThroughputExceeded

WriteProvisionedThroughputExceeded

ScalingType

data ScalingType

Constructors

UniformScaling

ShardIteratorType

data ShardIteratorType

Constructors

AfterSequenceNumber

AtSequenceNumber

AtTimestamp

Latest

TrimHorizon

StreamStatus

data StreamStatus

Constructors

Active

Creating

Deleting

Updating

EnhancedMetrics

data EnhancedMetrics

Represents enhanced metrics types.

See: enhancedMetrics smart constructor.

enhancedMetrics :: EnhancedMetrics

Creates a value of EnhancedMetrics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

emShardLevelMetrics - List of shard-level metrics. The following are the valid shard-level metrics. The value "ALL " enhances every metric. * IncomingBytes * IncomingRecords * OutgoingBytes * OutgoingRecords * WriteProvisionedThroughputExceeded * ReadProvisionedThroughputExceeded * IteratorAgeMilliseconds * ALL For more information, see Monitoring the Amazon Kinesis Streams Service with Amazon CloudWatch in the Amazon Kinesis Streams Developer Guide .

emShardLevelMetrics :: Lens' EnhancedMetrics [MetricsName]

List of shard-level metrics. The following are the valid shard-level metrics. The value "ALL " enhances every metric. * IncomingBytes * IncomingRecords * OutgoingBytes * OutgoingRecords * WriteProvisionedThroughputExceeded * ReadProvisionedThroughputExceeded * IteratorAgeMilliseconds * ALL For more information, see Monitoring the Amazon Kinesis Streams Service with Amazon CloudWatch in the Amazon Kinesis Streams Developer Guide .

EnhancedMonitoringOutput

data EnhancedMonitoringOutput

Represents the output for EnableEnhancedMonitoring and DisableEnhancedMonitoring .

See: enhancedMonitoringOutput smart constructor.

enhancedMonitoringOutput :: EnhancedMonitoringOutput

Creates a value of EnhancedMonitoringOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

emoDesiredShardLevelMetrics - Represents the list of all the metrics that would be in the enhanced state after the operation.

emoCurrentShardLevelMetrics - Represents the current state of the metrics that are in the enhanced state before the operation.

emoStreamName - The name of the Amazon Kinesis stream.

emoDesiredShardLevelMetrics :: Lens' EnhancedMonitoringOutput [MetricsName]

Represents the list of all the metrics that would be in the enhanced state after the operation.

emoCurrentShardLevelMetrics :: Lens' EnhancedMonitoringOutput [MetricsName]

Represents the current state of the metrics that are in the enhanced state before the operation.

emoStreamName :: Lens' EnhancedMonitoringOutput (Maybe Text)

The name of the Amazon Kinesis stream.

HashKeyRange

data HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

See: hashKeyRange smart constructor.

hashKeyRange

Arguments

:: Text
hkrStartingHashKey

-> Text
hkrEndingHashKey

-> HashKeyRange

Creates a value of HashKeyRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hkrStartingHashKey - The starting hash key of the hash key range.

hkrEndingHashKey - The ending hash key of the hash key range.

hkrStartingHashKey :: Lens' HashKeyRange Text

The starting hash key of the hash key range.

hkrEndingHashKey :: Lens' HashKeyRange Text

The ending hash key of the hash key range.

PutRecordsRequestEntry

data PutRecordsRequestEntry

Represents the output for PutRecords .

See: putRecordsRequestEntry smart constructor.

putRecordsRequestEntry

Arguments

:: ByteString
prreData

-> Text
prrePartitionKey

-> PutRecordsRequestEntry

Creates a value of PutRecordsRequestEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prreExplicitHashKey - The hash value used to determine explicitly the shard that the data record is assigned to by overriding the partition key hash.

prreData - The data blob to put into the record, which is base64-encoded when the blob is serialized. When the data blob (the payload before base64-encoding) is added to the partition key size, the total size must not exceed the maximum record size (1 MB).-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

prrePartitionKey - Determines which shard in the stream the data record is assigned to. Partition keys are Unicode strings with a maximum length limit of 256 characters for each key. Amazon Kinesis uses the partition key as input to a hash function that maps the partition key and associated data to a specific shard. Specifically, an MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key map to the same shard within the stream.

prreExplicitHashKey :: Lens' PutRecordsRequestEntry (Maybe Text)

The hash value used to determine explicitly the shard that the data record is assigned to by overriding the partition key hash.

prreData :: Lens' PutRecordsRequestEntry ByteString

The data blob to put into the record, which is base64-encoded when the blob is serialized. When the data blob (the payload before base64-encoding) is added to the partition key size, the total size must not exceed the maximum record size (1 MB).-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

prrePartitionKey :: Lens' PutRecordsRequestEntry Text

Determines which shard in the stream the data record is assigned to. Partition keys are Unicode strings with a maximum length limit of 256 characters for each key. Amazon Kinesis uses the partition key as input to a hash function that maps the partition key and associated data to a specific shard. Specifically, an MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key map to the same shard within the stream.

PutRecordsResultEntry

data PutRecordsResultEntry

Represents the result of an individual record from a PutRecords request. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to the stream includes ErrorCode and ErrorMessage in the result.

See: putRecordsResultEntry smart constructor.

putRecordsResultEntry :: PutRecordsResultEntry

Creates a value of PutRecordsResultEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prreSequenceNumber - The sequence number for an individual record result.

prreErrorCode - The error code for an individual record result. ErrorCodes can be either ProvisionedThroughputExceededException or InternalFailure .

prreErrorMessage - The error message for an individual record result. An ErrorCode value of ProvisionedThroughputExceededException has an error message that includes the account ID, stream name, and shard ID. An ErrorCode value of InternalFailure has the error message "Internal Service Failure" .

prreShardId - The shard ID for an individual record result.

prreSequenceNumber :: Lens' PutRecordsResultEntry (Maybe Text)

The sequence number for an individual record result.

prreErrorCode :: Lens' PutRecordsResultEntry (Maybe Text)

The error code for an individual record result. ErrorCodes can be either ProvisionedThroughputExceededException or InternalFailure .

prreErrorMessage :: Lens' PutRecordsResultEntry (Maybe Text)

The error message for an individual record result. An ErrorCode value of ProvisionedThroughputExceededException has an error message that includes the account ID, stream name, and shard ID. An ErrorCode value of InternalFailure has the error message "Internal Service Failure" .

prreShardId :: Lens' PutRecordsResultEntry (Maybe Text)

The shard ID for an individual record result.

Record

data Record

The unit of data of the Amazon Kinesis stream, which is composed of a sequence number, a partition key, and a data blob.

See: record smart constructor.

record

Arguments

:: Text
rSequenceNumber

-> ByteString
rData

-> Text
rPartitionKey

-> Record

Creates a value of Record with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rEncryptionType - The encryption type used on the record. This parameter can be one of the following values: * NONE : Do not encrypt the records in the stream. * KMS : Use server-side encryption on the records in the stream using a customer-managed KMS key.

rApproximateArrivalTimestamp - The approximate time that the record was inserted into the stream.

rSequenceNumber - The unique identifier of the record within its shard.

rData - The data blob. The data in the blob is both opaque and immutable to the Amazon Kinesis service, which does not inspect, interpret, or change the data in the blob in any way. When the data blob (the payload before base64-encoding) is added to the partition key size, the total size must not exceed the maximum record size (1 MB).-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

rPartitionKey - Identifies which shard in the stream the data record is assigned to.

rEncryptionType :: Lens' Record (Maybe EncryptionType)

The encryption type used on the record. This parameter can be one of the following values: * NONE : Do not encrypt the records in the stream. * KMS : Use server-side encryption on the records in the stream using a customer-managed KMS key.

rApproximateArrivalTimestamp :: Lens' Record (Maybe UTCTime)

The approximate time that the record was inserted into the stream.

rSequenceNumber :: Lens' Record Text

The unique identifier of the record within its shard.

rData :: Lens' Record ByteString

The data blob. The data in the blob is both opaque and immutable to the Amazon Kinesis service, which does not inspect, interpret, or change the data in the blob in any way. When the data blob (the payload before base64-encoding) is added to the partition key size, the total size must not exceed the maximum record size (1 MB).-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

rPartitionKey :: Lens' Record Text

Identifies which shard in the stream the data record is assigned to.

SequenceNumberRange

data SequenceNumberRange

The range of possible sequence numbers for the shard.

See: sequenceNumberRange smart constructor.

sequenceNumberRange

Arguments

:: Text
snrStartingSequenceNumber

-> SequenceNumberRange

Creates a value of SequenceNumberRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

snrEndingSequenceNumber - The ending sequence number for the range. Shards that are in the OPEN state have an ending sequence number of null .

snrStartingSequenceNumber - The starting sequence number for the range.

snrEndingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The ending sequence number for the range. Shards that are in the OPEN state have an ending sequence number of null .

snrStartingSequenceNumber :: Lens' SequenceNumberRange Text

The starting sequence number for the range.

Shard

data Shard

A uniquely identified group of data records in an Amazon Kinesis stream.

See: shard smart constructor.

shard

Arguments

:: Text
sShardId

-> HashKeyRange
sHashKeyRange

-> SequenceNumberRange
sSequenceNumberRange

-> Shard

Creates a value of Shard with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sAdjacentParentShardId - The shard ID of the shard adjacent to the shard's parent.

sParentShardId - The shard ID of the shard's parent.

sShardId - The unique identifier of the shard within the stream.

sHashKeyRange - The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

sSequenceNumberRange - The range of possible sequence numbers for the shard.

sAdjacentParentShardId :: Lens' Shard (Maybe Text)

The shard ID of the shard adjacent to the shard's parent.

sParentShardId :: Lens' Shard (Maybe Text)

The shard ID of the shard's parent.

sShardId :: Lens' Shard Text

The unique identifier of the shard within the stream.

sHashKeyRange :: Lens' Shard HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

sSequenceNumberRange :: Lens' Shard SequenceNumberRange

The range of possible sequence numbers for the shard.

StreamDescription

data StreamDescription

Represents the output for DescribeStream .

See: streamDescription smart constructor.

streamDescription

Arguments

:: Text
sdStreamName

-> Text
sdStreamARN

-> StreamStatus
sdStreamStatus

-> Bool
sdHasMoreShards

-> Natural
sdRetentionPeriodHours

-> UTCTime
sdStreamCreationTimestamp

-> StreamDescription

Creates a value of StreamDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdEncryptionType - The server-side encryption type used on the stream. This parameter can be one of the following values: * NONE : Do not encrypt the records in the stream. * KMS : Use server-side encryption on the records in the stream using a customer-managed KMS key.

sdKeyId - The GUID for the customer-managed KMS key used for encryption on the stream.

sdStreamName - The name of the stream being described.

sdStreamARN - The Amazon Resource Name (ARN) for the stream being described.

sdStreamStatus - The current status of the stream being described. The stream status is one of the following states: * CREATING - The stream is being created. Amazon Kinesis immediately returns and sets StreamStatus to CREATING . * DELETING - The stream is being deleted. The specified stream is in the DELETING state until Amazon Kinesis completes the deletion. * ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. * UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

sdShards - The shards that comprise the stream.

sdHasMoreShards - If set to true , more shards in the stream are available to describe.

sdRetentionPeriodHours - The current retention period, in hours.

sdStreamCreationTimestamp - The approximate time that the stream was created.

sdEnhancedMonitoring - Represents the current enhanced monitoring settings of the stream.

sdEncryptionType :: Lens' StreamDescription (Maybe EncryptionType)

The server-side encryption type used on the stream. This parameter can be one of the following values: * NONE : Do not encrypt the records in the stream. * KMS : Use server-side encryption on the records in the stream using a customer-managed KMS key.

sdKeyId :: Lens' StreamDescription (Maybe Text)

The GUID for the customer-managed KMS key used for encryption on the stream.

sdStreamName :: Lens' StreamDescription Text

The name of the stream being described.

sdStreamARN :: Lens' StreamDescription Text

The Amazon Resource Name (ARN) for the stream being described.

sdStreamStatus :: Lens' StreamDescription StreamStatus

The current status of the stream being described. The stream status is one of the following states: * CREATING - The stream is being created. Amazon Kinesis immediately returns and sets StreamStatus to CREATING . * DELETING - The stream is being deleted. The specified stream is in the DELETING state until Amazon Kinesis completes the deletion. * ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. * UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

sdShards :: Lens' StreamDescription [Shard]

The shards that comprise the stream.

sdHasMoreShards :: Lens' StreamDescription Bool

If set to true , more shards in the stream are available to describe.

sdRetentionPeriodHours :: Lens' StreamDescription Natural

The current retention period, in hours.

sdStreamCreationTimestamp :: Lens' StreamDescription UTCTime

The approximate time that the stream was created.

sdEnhancedMonitoring :: Lens' StreamDescription [EnhancedMetrics]

Represents the current enhanced monitoring settings of the stream.

Tag

data Tag

Metadata assigned to the stream, consisting of a key-value pair.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - An optional string, typically used to describe or define the tag. Maximum length: 256 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @

tagKey - A unique identifier for the tag. Maximum length: 128 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @

tagValue :: Lens' Tag (Maybe Text)

An optional string, typically used to describe or define the tag. Maximum length: 256 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @

tagKey :: Lens' Tag Text

A unique identifier for the tag. Maximum length: 128 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @