Network.AWS.Kinesis.Types

Service

data Kinesis

Version 2013-12-02 of the Amazon Kinesis service.

Error

data JSONError :: *

Shard

data Shard

shard

Arguments

:: Text
sShardId

-> HashKeyRange
sHashKeyRange

-> SequenceNumberRange
sSequenceNumberRange

-> Shard

Shard constructor.

The fields accessible through corresponding lenses are:

sAdjacentParentShardId :: Maybe Text

sHashKeyRange :: HashKeyRange

sParentShardId :: Maybe Text

sSequenceNumberRange :: SequenceNumberRange

sShardId :: Text

sAdjacentParentShardId :: Lens' Shard (Maybe Text)

The shard Id of the shard adjacent to the shard's parent.

sHashKeyRange :: Lens' Shard HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

sParentShardId :: Lens' Shard (Maybe Text)

The shard Id of the shard's parent.

sSequenceNumberRange :: Lens' Shard SequenceNumberRange

The range of possible sequence numbers for the shard.

sShardId :: Lens' Shard Text

The unique identifier of the shard within the Amazon Kinesis stream.

Tag

data Tag

tag

Arguments

:: Text
tagKey

-> Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Text

tagValue :: Maybe Text

tagKey :: Lens' Tag Text

A unique identifier for the tag. Maximum length: 128 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @.

tagValue :: Lens' Tag (Maybe Text)

An optional string, typically used to describe or define the tag. Maximum length: 256 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @.

StreamDescription

data StreamDescription

streamDescription

Arguments

:: Text
sdStreamName

-> Text
sdStreamARN

-> Text
sdStreamStatus

-> Bool
sdHasMoreShards

-> StreamDescription

StreamDescription constructor.

The fields accessible through corresponding lenses are:

sdHasMoreShards :: Bool

sdShards :: [Shard]

sdStreamARN :: Text

sdStreamName :: Text

sdStreamStatus :: Text

sdHasMoreShards :: Lens' StreamDescription Bool

If set to true, more shards in the stream are available to describe.

sdShards :: Lens' StreamDescription [Shard]

The shards that comprise the stream.

sdStreamARN :: Lens' StreamDescription Text

The Amazon Resource Name (ARN) for the stream being described.

sdStreamName :: Lens' StreamDescription Text

The name of the stream being described.

sdStreamStatus :: Lens' StreamDescription Text

The current status of the stream being described. The stream status is one of the following states: CREATING - The stream is being created. Amazon Kinesis immediately returns and sets StreamStatus to CREATING. DELETING - The stream is being deleted. The specified stream is in the DELETING state until Amazon Kinesis completes the deletion. ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

StreamStatus

data StreamStatus

Constructors

Active
ACTIVE

Creating
CREATING

Deleting
DELETING

Updating
UPDATING

HashKeyRange

data HashKeyRange

hashKeyRange

Arguments

:: Text
hkrStartingHashKey

-> Text
hkrEndingHashKey

-> HashKeyRange

HashKeyRange constructor.

The fields accessible through corresponding lenses are:

hkrEndingHashKey :: Text

hkrStartingHashKey :: Text

hkrEndingHashKey :: Lens' HashKeyRange Text

The ending hash key of the hash key range.

hkrStartingHashKey :: Lens' HashKeyRange Text

The starting hash key of the hash key range.

Record

data Record

record

Arguments

:: Text
rSequenceNumber

-> Base64
rData

-> Text
rPartitionKey

-> Record

Record constructor.

The fields accessible through corresponding lenses are:

rData :: Base64

rPartitionKey :: Text

rSequenceNumber :: Text

rData :: Lens' Record Base64

The data blob. The data in the blob is both opaque and immutable to the Amazon Kinesis service, which does not inspect, interpret, or change the data in the blob in any way. The maximum size of the data blob (the payload before Base64-encoding) is 50 kilobytes (KB).

rPartitionKey :: Lens' Record Text

Identifies which shard in the stream the data record is assigned to.

rSequenceNumber :: Lens' Record Text

The unique identifier for the record in the Amazon Kinesis stream.

SequenceNumberRange

data SequenceNumberRange

sequenceNumberRange

Arguments

:: Text
snrStartingSequenceNumber

-> SequenceNumberRange

SequenceNumberRange constructor.

The fields accessible through corresponding lenses are:

snrEndingSequenceNumber :: Maybe Text

snrStartingSequenceNumber :: Text

snrEndingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The ending sequence number for the range. Shards that are in the OPEN state have an ending sequence number of null.

snrStartingSequenceNumber :: Lens' SequenceNumberRange Text

The starting sequence number for the range.

ShardIteratorType

data ShardIteratorType

Constructors

AfterSequenceNumber
AFTER_SEQUENCE_NUMBER

AtSequenceNumber
AT_SEQUENCE_NUMBER

Latest
LATEST

TrimHorizon
TRIM_HORIZON