Network.AWS.Kinesis.Types

Service

data Kinesis

Version 2013-12-02 of the Amazon Kinesis service.

Error

data JSONError :: *

Shard

data Shard

shard

Arguments

:: Text
sShardId

-> HashKeyRange
sHashKeyRange

-> SequenceNumberRange
sSequenceNumberRange

-> Shard

Shard constructor.

The fields accessible through corresponding lenses are:

sAdjacentParentShardId :: Maybe Text

sHashKeyRange :: HashKeyRange

sParentShardId :: Maybe Text

sSequenceNumberRange :: SequenceNumberRange

sShardId :: Text

sAdjacentParentShardId :: Lens' Shard (Maybe Text)

The shard Id of the shard adjacent to the shard's parent.

sHashKeyRange :: Lens' Shard HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered contiguous positive integers.

sParentShardId :: Lens' Shard (Maybe Text)

The shard Id of the shard's parent.

sSequenceNumberRange :: Lens' Shard SequenceNumberRange

The range of possible sequence numbers for the shard.

sShardId :: Lens' Shard Text

The unique identifier of the shard within the Amazon Kinesis stream.

Tag

data Tag

tag

Arguments

:: Text
tagKey

-> Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Text

tagValue :: Maybe Text

tagKey :: Lens' Tag Text

A unique identifier for the tag. Maximum length: 128 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @

tagValue :: Lens' Tag (Maybe Text)

An optional string, typically used to describe or define the tag. Maximum length: 256 characters. Valid characters: Unicode letters, digits, white space, _ . / = + - % @

StreamDescription

data StreamDescription

streamDescription

Arguments

:: Text
sdStreamName

-> Text
sdStreamARN

-> StreamStatus
sdStreamStatus

-> Bool
sdHasMoreShards

-> StreamDescription

StreamDescription constructor.

The fields accessible through corresponding lenses are:

sdHasMoreShards :: Bool

sdShards :: [Shard]

sdStreamARN :: Text

sdStreamName :: Text

sdStreamStatus :: StreamStatus

sdHasMoreShards :: Lens' StreamDescription Bool

If set to true, more shards in the stream are available to describe.

sdShards :: Lens' StreamDescription [Shard]

The shards that comprise the stream.

sdStreamARN :: Lens' StreamDescription Text

The Amazon Resource Name (ARN) for the stream being described.

sdStreamName :: Lens' StreamDescription Text

The name of the stream being described.

sdStreamStatus :: Lens' StreamDescription StreamStatus

The current status of the stream being described.

The stream status is one of the following states:

CREATING - The stream is being created. Amazon Kinesis immediately returns and sets StreamStatus to CREATING. DELETING - The stream is being deleted. The specified stream is in the DELETING state until Amazon Kinesis completes the deletion. ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream. UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

PutRecordsResultEntry

data PutRecordsResultEntry

putRecordsResultEntry :: PutRecordsResultEntry

PutRecordsResultEntry constructor.

The fields accessible through corresponding lenses are:

prreErrorCode :: Maybe Text

prreErrorMessage :: Maybe Text

prreSequenceNumber :: Maybe Text

prreShardId :: Maybe Text

prreErrorCode :: Lens' PutRecordsResultEntry (Maybe Text)

The error code for an individual record result. ErrorCodes can be either ProvisionedThroughputExceededException or InternalFailure.

prreErrorMessage :: Lens' PutRecordsResultEntry (Maybe Text)

The error message for an individual record result. An ErrorCode value of ProvisionedThroughputExceededException has an error message that includes the account ID, stream name, and shard ID. An ErrorCode value of InternalFailure has the error message '"InternalService Failure"'.

prreSequenceNumber :: Lens' PutRecordsResultEntry (Maybe Text)

The sequence number for an individual record result.

prreShardId :: Lens' PutRecordsResultEntry (Maybe Text)

The shard ID for an individual record result.

PutRecordsRequestEntry

data PutRecordsRequestEntry

putRecordsRequestEntry

Arguments

:: Base64
prreData

-> Text
prrePartitionKey

-> PutRecordsRequestEntry

PutRecordsRequestEntry constructor.

The fields accessible through corresponding lenses are:

prreData :: Base64

prreExplicitHashKey :: Maybe Text

prrePartitionKey :: Text

prreData :: Lens' PutRecordsRequestEntry Base64

The data blob to put into the record, which is base64-encoded when the blob is serialized. The maximum size of the data blob (the payload before base64-encoding) is 50 kilobytes (KB)

prreExplicitHashKey :: Lens' PutRecordsRequestEntry (Maybe Text)

The hash value used to determine explicitly the shard that the data record is assigned to by overriding the partition key hash.

prrePartitionKey :: Lens' PutRecordsRequestEntry Text

Determines which shard in the stream the data record is assigned to. Partition keys are Unicode strings with a maximum length limit of 256 bytes. Amazon Kinesis uses the partition key as input to a hash function that maps the partition key and associated data to a specific shard. Specifically, an MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key map to the same shard within the stream.

StreamStatus

data StreamStatus

Constructors

Active
ACTIVE

Creating
CREATING

Deleting
DELETING

Updating
UPDATING

HashKeyRange

data HashKeyRange

hashKeyRange

Arguments

:: Text
hkrStartingHashKey

-> Text
hkrEndingHashKey

-> HashKeyRange

HashKeyRange constructor.

The fields accessible through corresponding lenses are:

hkrEndingHashKey :: Text

hkrStartingHashKey :: Text

hkrEndingHashKey :: Lens' HashKeyRange Text

The ending hash key of the hash key range.

hkrStartingHashKey :: Lens' HashKeyRange Text

The starting hash key of the hash key range.

Record

data Record

record

Arguments

:: Text
rSequenceNumber

-> Base64
rData

-> Text
rPartitionKey

-> Record

Record constructor.

The fields accessible through corresponding lenses are:

rData :: Base64

rPartitionKey :: Text

rSequenceNumber :: Text

rData :: Lens' Record Base64

The data blob. The data in the blob is both opaque and immutable to the Amazon Kinesis service, which does not inspect, interpret, or change the data in the blob in any way. The maximum size of the data blob (the payload before base64-encoding) is 50 kilobytes (KB)

rPartitionKey :: Lens' Record Text

Identifies which shard in the stream the data record is assigned to.

rSequenceNumber :: Lens' Record Text

The unique identifier for the record in the Amazon Kinesis stream.

SequenceNumberRange

data SequenceNumberRange

sequenceNumberRange

Arguments

:: Text
snrStartingSequenceNumber

-> SequenceNumberRange

SequenceNumberRange constructor.

The fields accessible through corresponding lenses are:

snrEndingSequenceNumber :: Maybe Text

snrStartingSequenceNumber :: Text

snrEndingSequenceNumber :: Lens' SequenceNumberRange (Maybe Text)

The ending sequence number for the range. Shards that are in the OPEN state have an ending sequence number of null.

snrStartingSequenceNumber :: Lens' SequenceNumberRange Text

The starting sequence number for the range.

ShardIteratorType

data ShardIteratorType

Constructors

AfterSequenceNumber
AFTER_SEQUENCE_NUMBER

AtSequenceNumber
AT_SEQUENCE_NUMBER

Latest
LATEST

TrimHorizon
TRIM_HORIZON